refactor: organize agent client/server tests into dedicated subfolder
Move agent-related test files from root __tests__ directory to __tests__/agent/:
- test-server.ts, test-server-auth.ts - Agent server implementations
- test-client.ts, test-client-auth.ts - Agent client implementations
- run-test.ts, run-auth-test.ts - Integration test runners
- python-client/ - Cross-language client tests

Update import paths and file references to maintain functionality.
Improves test organization by grouping related agent functionality.
